requirements.txt
Django==1.9.6
pillow==3.2.0
django-taggit==0.19.1
djangoajax==2.3.6
python-social-auth==0.2.19
psycopg2==2.6.1